[FEAT/#37] 남은 시간 계산 함수 구현

+++ domain/src/main/java/com/example/domain/entity/type/YelloState.kt
@@ -3,5 +3,5 @@ package com.example.domain.entity.type
 sealed class YelloState {
     object Lock : YelloState()
     data class Valid(val point: Int) : YelloState()
-    data class Wait(val leftSec: Int) : YelloState()
+    data class Wait(val leftSec: Long) : YelloState()
 }
