+++ domain/src/main/java/com/example/domain/entity/vote/VoteState.kt
@@ -3,5 +3,5 @@ package com.example.domain.entity.vote
 data class VoteState(
     val isStart: Boolean,
     val point: Int,
-    val leftTime: Int,
+    val leftTime: Long,
 )
